// src/app/web/api/position.ts
export interface Position {

    position_id: number;
    position_parent_id: number;
    org_id: number;
    div_id: number;
    org_name: string;
    division_name: string;
    position_name: string;
    position_name1: string;
    position_type: string;
    lov_id:string;
    position_status: string;
    createdDate: string;
    createdBy: string;
    updatedDate: string;
    updatedBy: string;
    approve_flag: string;
  }